# Remove alloc overhead for `NormalizedValue`