Was ~200KB/s during gameplay.

Not even going to provide a benchmark, should be obvious.