@bdach 2024-01-09
Just for my peace of mind...

before

|                      Method |     Mean |    Error |   StdDev |   Gen0 | Allocated |
|---------------------------- |---------:|---------:|---------:|-------:|----------:|
| NormalizedValueDoesNotAlloc | 27.65 us | 0.472 us | 0.580 us | 5.7068 |  70.24 KB |

after

|                      Method |     Mean |     Error |    StdDev | Allocated |
|---------------------------- |---------:|----------:|----------:|----------:|
| NormalizedValueDoesNotAlloc | 5.264 us | 0.0489 us | 0.0409 us |         - |

Not committing the benchmark though, was very ad-hoc and I don't care enough to clean it up.